requirements.txt
# POS System Requirements

Pillow>=9.0.0